Current code upload
Everything is functional, manuals haven't been implemented yet.

The XML file vCommands/Properties/Default Manual.xml will hold all the
manuals for the default commands in the future. The test content in
there is, well, for testing.

Note that the assembly is signed with the key
vCommands/Properties/vercas_spc.pfx which is obviously not present.